Improvements to the context
