+++ src/main/kotlin/eu/europa/ec/eudi/pidissuer/IssuerContext.kt
@@ -15,37 +15,20 @@
  */
 package eu.europa.ec.eudi.pidissuer
 
-import arrow.core.nonEmptySetOf
-import arrow.core.recover
-import arrow.core.some
 import arrow.core.toNonEmptyListOrNull
-import com.nimbusds.jose.JWSAlgorithm
-import com.nimbusds.jose.jwk.Curve
-import com.nimbusds.jose.jwk.ECKey
-import com.nimbusds.jose.jwk.KeyUse
-import com.nimbusds.jose.jwk.gen.ECKeyGenerator
-import com.nimbusds.jose.util.Base64
-import com.nimbusds.oauth2.sdk.id.Issuer
-import com.nimbusds.oauth2.sdk.util.X509CertificateUtils
 import eu.europa.ec.eudi.pidissuer.adapter.input.scheduler.CredentialRevocationJob
 import eu.europa.ec.eudi.pidissuer.adapter.input.web.*
-import eu.europa.ec.eudi.pidissuer.adapter.input.web.security.DPoPConfigurationProperties
-import eu.europa.ec.eudi.pidissuer.adapter.out.IssuerSigningKey
-import eu.europa.ec.eudi.pidissuer.adapter.out.jose.AccessCertificate
 import eu.europa.ec.eudi.pidissuer.adapter.out.jose.EncryptCredentialResponseNimbus
 import eu.europa.ec.eudi.pidissuer.adapter.out.jose.EncryptDeferredResponseNimbus
 import eu.europa.ec.eudi.pidissuer.adapter.out.jose.GenerateSignedMetadataWithNimbus
 import eu.europa.ec.eudi.pidissuer.adapter.out.nonce.DecryptNonceWithNimbusAndVerify
 import eu.europa.ec.eudi.pidissuer.adapter.out.nonce.GenerateNonceAndEncryptWithNimbus
-import eu.europa.ec.eudi.pidissuer.adapter.out.nonce.NonceEncryptionKey
 import eu.europa.ec.eudi.pidissuer.adapter.out.persistence.InMemoryDeferredCredentialRepository
 import eu.europa.ec.eudi.pidissuer.adapter.out.persistence.R2dbcIssuedCredentialRepository
 import eu.europa.ec.eudi.pidissuer.adapter.out.qr.DefaultGenerateQrCode
 import eu.europa.ec.eudi.pidissuer.adapter.out.status.GenerateStatusListTokenWithExternalService
 import eu.europa.ec.eudi.pidissuer.adapter.out.status.GetStatusListTokenWithStatium
 import eu.europa.ec.eudi.pidissuer.adapter.out.status.MarkStatusAsRevokedWithExternalService
-import eu.europa.ec.eudi.pidissuer.adapter.out.trust.Ignored
-import eu.europa.ec.eudi.pidissuer.adapter.out.trust.usingTrustValidatorService
 import eu.europa.ec.eudi.pidissuer.adapter.out.webclient.KtorHttpClients
 import eu.europa.ec.eudi.pidissuer.adapter.out.webclient.WebClients
 import eu.europa.ec.eudi.pidissuer.domain.*
@@ -56,7 +39,6 @@ import eu.europa.ec.eudi.pidissuer.port.out.persistence.GenerateTransactionId
 import eu.europa.ec.eudi.pidissuer.port.out.status.AllocateStatus
 import eu.europa.ec.eudi.pidissuer.port.out.status.GetStatusListTokenStatus
 import eu.europa.ec.eudi.pidissuer.port.out.status.MarkStatusAsRevoked
-import eu.europa.ec.eudi.pidissuer.port.out.trust.IsTrustedKeyAttestationIssuer
 import eu.europa.ec.eudi.sdjwt.vc.Vct
 import kotlinx.coroutines.runBlocking
 import kotlinx.datetime.TimeZone
@@ -64,10 +46,6 @@ import kotlinx.serialization.json.Json
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.BeanRegistrarDsl
 import org.springframework.boot.http.codec.CodecCustomizer
-import org.springframework.core.env.Environment
-import org.springframework.core.env.getProperty
-import org.springframework.core.io.DefaultResourceLoader
-import org.springframework.core.io.FileSystemResource
 import org.springframework.http.codec.json.KotlinSerializationJsonDecoder
 import org.springframework.http.codec.json.KotlinSerializationJsonEncoder
 import org.springframework.scheduling.annotation.SchedulingConfigurer
@@ -76,14 +54,11 @@ import java.net.URI
 import java.security.KeyStore
 import java.util.*
 import kotlin.time.Clock
-import kotlin.time.Duration.Companion.days
 import kotlin.time.Duration.Companion.minutes
 import kotlin.time.Duration.Companion.seconds
 import kotlin.time.toKotlinDuration
 
 val log = LoggerFactory.getLogger(PidIssuerApplication::class.java)
-
-private const val KEYSTORE_DEFAULT_LOCATION = "/keystore.jks"
 
 fun beans(
     clock: Clock,
@@ -91,114 +66,22 @@ fun beans(
 ) = BeanRegistrarDsl {
     val issuerPublicUrl = env.readRequiredUrl("issuer.publicUrl", removeTrailingSlash = true)
     val credentialsOfferUri = env.getRequiredProperty("issuer.credentialOffer.uri")
-    val trustValidatorServiceUrl = env.getProperty<String>("issuer.trust.service-url")
-    val issuerKeystore: KeyStore by lazy {
-        val keystoreLocation = env.getRequiredProperty("issuer.keystore.file")
-        log.info("Will try to load Keystore from: '{}'", keystoreLocation)
-        val keystoreResource =
-            DefaultResourceLoader()
-                .getResource(keystoreLocation)
-                .some()
-                .filter { it.exists() }
-                .recover {
-                    log.warn(
-                        "Could not find Keystore at '{}'. Fallback to '{}'",
-                        keystoreLocation,
-                        KEYSTORE_DEFAULT_LOCATION,
-                    )
-                    FileSystemResource(KEYSTORE_DEFAULT_LOCATION)
-                        .some()
-                        .filter { it.exists() }
-                        .bind()
-                }.getOrNull()
-        checkNotNull(keystoreResource) { "Could not load Keystore either from '$keystoreLocation' or '$KEYSTORE_DEFAULT_LOCATION'" }
+    val issuerKeystore: KeyStore by lazy { keystore(env) }
+    val getIssuerSigningKey = loadOrGenerateIssuerSigningKey(clock, env, issuerPublicUrl) { issuerKeystore }
+    val proxy = env.httpProxy()
 
-        val keystoreType = env.getProperty("issuer.keystore.type", KeyStore.getDefaultType())
-        val keystorePassword = env.getProperty("issuer.keystore.password")?.takeIf { it.isNotBlank() }
-
-        keystoreResource.inputStream.use { inputStream ->
-            val keystore = KeyStore.getInstance(keystoreType)
-            keystore.load(inputStream, keystorePassword?.toCharArray())
-            keystore
-        }
-    }
-
-    fun getIssuerSigningKey(prefix: String): IssuerSigningKey {
-        val signingKey =
-            when (env.getProperty<KeyOption>(prefix)) {
-                null, KeyOption.GenerateRandom -> {
-                    log.info("Generating random signing key and self-signed certificate for '$prefix'")
-                    val key = ECKeyGenerator(Curve.P_256).keyID("issuer-kid-$prefix").generate()
-                    val certificate =
-                        X509CertificateUtils.generateSelfSigned(
-                            Issuer(issuerPublicUrl.value.host),
-                            clock.now().toJavaDate(),
-                            (clock.now() + 365.days).toJavaDate(),
-                            key.toECPublicKey(),
-                            key.toECPrivateKey(),
-                        )
-                    ECKey
-                        .Builder(key)
-                        .x509CertChain(listOf(Base64.encode(certificate.encoded)))
-                        .build()
-                }
-
-                KeyOption.LoadFromKeystore -> {
-                    log.info("Loading signing key and certificate for issuance from keystore for '$prefix'")
-                    issuerKeystore.loadJwk(env, prefix)
-                }
-            }
-        require(signingKey is ECKey) { "Only ECKeys are supported for signing" }
-        return IssuerSigningKey(signingKey)
-    }
-
-    fun accessCertificate(): AccessCertificate {
-        val key =
-            when (env.getProperty<KeyOption>("issuer.access-certificate")) {
-                null, KeyOption.GenerateRandom -> {
-                    log.info("Generating random access certificate key for metadata signing")
-                    ECKeyGenerator(Curve.P_256)
-                        .keyID("issuer-kid-1")
-                        .keyUse(KeyUse.SIGNATURE)
-                        .generate()
-                }
-
-                KeyOption.LoadFromKeystore -> {
-                    log.info("Loading access certificate for metadata signing from keystore")
-                    issuerKeystore.loadJwk(env, "issuer.access-certificate")
-                }
-            }
-
-        return AccessCertificate(key)
-    }
-
-    //
-    // Nonce encryption key
-    //
-    val nonceEncryptionKey = loadNonceEncryptionKey(env) { issuerKeystore }
-
-    //
-    // Signed metadata signing key
-    //
-    registerBean(lazyInit = true) { accessCertificate() }
-
-    //
-    // Adapters (out ports)
-    //
+    registerBean { env.dPoPConfigurationProperties() }
+    registerBean(lazyInit = true) { loadOrGenerateAccessCertificate(env) { issuerKeystore } }
     registerBean { clock }
     registerBean { timeZone }
-
-    val proxy = httpProxy(env)
-    val webClient = WebClients(proxy, secure = "insecure" !in env.activeProfiles)
-    registerBean { webClient }
-
+    registerBean { WebClients(proxy, secure = "insecure" !in env.activeProfiles) }
+    registerBean { trustValidatorService(env, bean()) }
     registerBean { DefaultGenerateQrCode() }
     registerBean { HandleNotificationRequest(bean()) }
     registerBean {
         val cNonceExpiresIn = env.duration("issuer.cnonce.expiration") ?: 5.minutes
         HandleNonceRequest(clock, cNonceExpiresIn, bean())
     }
-
     registerBean(lazyInit = true) {
         val signedMetadataIssuer =
             env
@@ -277,8 +160,9 @@ fun beans(
     //
     // Nonce
     //
-    registerBean { GenerateNonceAndEncryptWithNimbus(issuerPublicUrl, nonceEncryptionKey) }
-    registerBean { DecryptNonceWithNimbusAndVerify(issuerPublicUrl, nonceEncryptionKey) }
+    registerBean { loadNonceEncryptionKey(env) { issuerKeystore } }
+    registerBean { GenerateNonceAndEncryptWithNimbus(issuerPublicUrl, bean()) }
+    registerBean { DecryptNonceWithNimbusAndVerify(issuerPublicUrl, bean()) }
 
     //
     // Credentials
@@ -303,15 +187,6 @@ fun beans(
     //
     // Specific Issuers
     //
-    registerBean {
-        if (trustValidatorServiceUrl.isNullOrBlank()) {
-            log.warn("Trust Validator Service has not been configured. Trusting all Wallet Providers.")
-            IsTrustedKeyAttestationIssuer.Ignored
-        } else {
-            log.info("Using Trust Validator Service '{}'", trustValidatorServiceUrl)
-            IsTrustedKeyAttestationIssuer.usingTrustValidatorService(bean(), URI.create(trustValidatorServiceUrl))
-        }
-    }
 
     registerBean {
         getPidDataFromKeyCloak(clock, timeZone, env, webClient = bean())
@@ -417,17 +292,10 @@ fun beans(
         CreateCredentialsOffer(bean(), credentialsOfferUri)
     }
 
-    val dPoPConfigurationProperties =
-        DPoPConfigurationProperties(
-            nonEmptySetOf(JWSAlgorithm.ES256, JWSAlgorithm.ES384, JWSAlgorithm.ES512),
-            env.getProperty("issuer.dpop.realm")?.takeIf { it.isNotBlank() },
-            env.getProperty<Boolean>("issuer.dpop.nonce.enabled") ?: true,
-        )
-
     registerBean {
         GetProtectedResourceMetadata(
             credentialIssuerMetadata = bean(),
-            dPoPConfigurationProperties = dPoPConfigurationProperties,
+            dPoPConfigurationProperties = bean(),
         )
     }
 
@@ -458,8 +326,8 @@ fun beans(
             http = bean(),
             metadata = bean(),
             oAuth2ResourceServerProperties = bean(),
-            dPoPConfigurationProperties = dPoPConfigurationProperties,
-            webClient = webClient,
+            dPoPConfigurationProperties = bean(),
+            webClient = bean(),
             env = env,
             clock = clock,
             verifyNonce = bean(),
@@ -484,36 +352,5 @@ fun beans(
             val maxInMemorySize = DataSize.parse(env.getProperty("spring.webflux.codecs.max-in-memory-size", "1MB"))
             it.defaultCodecs().maxInMemorySize(maxInMemorySize.toBytes().toInt())
         }
-    }
-}
-
-private fun loadNonceEncryptionKey(
-    env: Environment,
-    issuerKeystore: () -> KeyStore,
-): NonceEncryptionKey {
-    val encryptionKey: ECKey =
-        when (env.getProperty<KeyOption>("issuer.nonce.encryption-key")) {
-            null, KeyOption.GenerateRandom -> {
-                log.info("Generating random encryption key for Nonce")
-                ECKeyGenerator(Curve.P_256).keyUse(KeyUse.ENCRYPTION).generate()
-            }
-
-            KeyOption.LoadFromKeystore -> {
-                log.info("Loading Nonce encryption key from keystore")
-                val nonceEncryptionKey = issuerKeystore().loadJwk(env, "issuer.nonce.encryption-key")
-                require(nonceEncryptionKey is ECKey) { "Only ECKey are supported for encryption" }
-                nonceEncryptionKey
-            }
-        }
-    return NonceEncryptionKey(encryptionKey)
-}
-
-private fun Environment.batchCredentialIssuance(): BatchCredentialIssuance {
-    val enabled = getProperty<Boolean>("issuer.credentialEndpoint.batchIssuance.enabled") ?: true
-    return if (enabled) {
-        val batchSize = getProperty<Int>("issuer.credentialEndpoint.batchIssuance.batchSize") ?: 10
-        BatchCredentialIssuance.Supported(batchSize)
-    } else {
-        BatchCredentialIssuance.NotSupported
     }
 }
